gpsd: use upstream config instead of our own
We have our own versions of the udev, hotplug, and systemd scripts,
but at this point they are several years old and were added to fix bugs
at the time. As far as I can tell, the relevant bugs are now fixed
upstream, so we can move back to the upstream versions.

Signed-off-by: Martin Kelly <mkelly@xevo.com>
Signed-off-by: Armin Kuster <akuster808@gmail.com>